Correccion al modulo area

+++ Restaurante/Restaurante/ModuloAreas.cs
@@ -135,7 +135,7 @@ namespace Restaurante
                 {
                     Clases.Restaurante.ModificarAreas
                         (
-                            Convert.ToInt16(this.id),
+                            Convert.ToInt32(this.id),
                             txtNombre.Text,
                             Convert.ToInt16(txtNumero.Text)
                         );
